[BIT-138] Moved all admin controller in a new folder
